+++ docs/html/search/all_2.js
@@ -1,7 +1,6 @@
 var searchData=
 [
-  ['getmaxlevel_5',['getMaxLevel',['../classace__segment_1_1LevelWriter.html#a7cb31a9ea04c9386bcf724dbc69ab5d0',1,'ace_segment::LevelWriter']]],
-  ['getnumchars_6',['getNumChars',['../classace__segment_1_1CharWriter.html#afeeba3cff35da1b673ba790da44c1d6d',1,'ace_segment::CharWriter']]],
-  ['getnumdigits_7',['getNumDigits',['../classace__segment_1_1CharWriter.html#a3dd53adff049e619eab49084a40da054',1,'ace_segment::CharWriter::getNumDigits()'],['../classace__segment_1_1PatternWriter.html#a8054a9ccd958a4e83fdc66b94ab9537d',1,'ace_segment::PatternWriter::getNumDigits()'],['../classace__segment_1_1StringScroller.html#ac70037cd50bfe1a80f6156cde68f51d4',1,'ace_segment::StringScroller::getNumDigits()']]],
-  ['getpattern_8',['getPattern',['../classace__segment_1_1CharWriter.html#a0252cd3e0a8b85c55737b3196c0b3196',1,'ace_segment::CharWriter']]]
+  ['getmaxlevel_7',['getMaxLevel',['../classace__segment_1_1LevelWriter.html#a7cb31a9ea04c9386bcf724dbc69ab5d0',1,'ace_segment::LevelWriter']]],
+  ['getnumchars_8',['getNumChars',['../classace__segment_1_1CharWriter.html#afeeba3cff35da1b673ba790da44c1d6d',1,'ace_segment::CharWriter']]],
+  ['getpattern_9',['getPattern',['../classace__segment_1_1CharWriter.html#a0252cd3e0a8b85c55737b3196c0b3196',1,'ace_segment::CharWriter']]]
 ];