+++ docs/html/search/all_8.js
@@ -1,4 +1,9 @@
 var searchData=
 [
-  ['temperaturewriter_19',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html',1,'ace_segment::TemperatureWriter&lt; T_LED_MODULE &gt;'],['../classace__segment_1_1TemperatureWriter.html#a581f8fabbe508e1c692c7342d5333f68',1,'ace_segment::TemperatureWriter::TemperatureWriter()']]]
+  ['scrollleft_20',['scrollLeft',['../classace__segment_1_1StringScroller.html#a633d294d5d38fccb85b52837d53b5622',1,'ace_segment::StringScroller']]],
+  ['scrollright_21',['scrollRight',['../classace__segment_1_1StringScroller.html#a4a4ac4f53ba4adc7a5411cbc4c47a42e',1,'ace_segment::StringScroller']]],
+  ['setdecimalpointat_22',['setDecimalPointAt',['../classace__segment_1_1CharWriter.html#aebec45ba87420c0dcf8704ca923d65eb',1,'ace_segment::CharWriter::setDecimalPointAt()'],['../classace__segment_1_1NumberWriter.html#a133423318fbde8f97e6290e759fd14ec',1,'ace_segment::NumberWriter::setDecimalPointAt()'],['../classace__segment_1_1PatternWriter.html#a14254837d740cea1bc6aa646c8e66264',1,'ace_segment::PatternWriter::setDecimalPointAt()']]],
+  ['size_23',['size',['../classace__segment_1_1CharWriter.html#ab8e6b85e023033c1ff7c975030db2321',1,'ace_segment::CharWriter::size()'],['../classace__segment_1_1PatternWriter.html#a7d466036b58f00803e45eb2ab522745a',1,'ace_segment::PatternWriter::size()'],['../classace__segment_1_1StringScroller.html#a89663d0277c4afd1df958512e921d742',1,'ace_segment::StringScroller::size()']]],
+  ['stringscroller_24',['StringScroller',['../classace__segment_1_1StringScroller.html',1,'ace_segment::StringScroller&lt; T_LED_MODULE &gt;'],['../classace__segment_1_1StringScroller.html#a3ffe6e513c6da0b4d822921e1b99a0eb',1,'ace_segment::StringScroller::StringScroller()']]],
+  ['stringwriter_25',['StringWriter',['../classace__segment_1_1StringWriter.html',1,'ace_segment::StringWriter&lt; T_LED_MODULE &gt;'],['../classace__segment_1_1StringWriter.html#a1db3ac016159e8ff0b8d5ca009c6f197',1,'ace_segment::StringWriter::StringWriter()']]]
 ];