+++ src/ace_segment_writer/LevelWriter.h
@@ -48,8 +48,8 @@ template <typename T_LED_MODULE>
 class LevelWriter {
   public:
     /** Constructor. */
-    explicit LevelWriter(T_LED_MODULE& ledModule) :
-        mPatternWriter(ledModule)
+    explicit LevelWriter(PatternWriter<T_LED_MODULE>& patternWriter) :
+        mPatternWriter(patternWriter)
     {}
 
     /** Get the underlying LedModule. */
@@ -63,24 +63,25 @@ class LevelWriter {
      * maxLevel] inclusive.
      */
     uint8_t getMaxLevel() const {
-      return mPatternWriter.getNumDigits() * 2;
+      return mPatternWriter.size() * 2;
     }
 
     /** Write out the level bar, 2 levels per digit. */
     void writeLevel(uint8_t level) {
       uint8_t fullDigits = level / 2;
       uint8_t partialDigit = level & 0x1;
-      uint8_t numDigits = mPatternWriter.getNumDigits();
+      uint8_t numDigits = mPatternWriter.size();
 
-      uint8_t pos = 0;
-      while (pos < fullDigits && pos < numDigits) {
-        mPatternWriter.writePatternAt(
-            pos++, kPatternLevelLeft | kPatternLevelRight);
+      mPatternWriter.home();
+      for (;;) {
+        uint8_t pos = mPatternWriter.pos();
+        if (pos >= fullDigits || pos >= numDigits) { break; }
+        mPatternWriter.writePattern(kPatternLevelLeft | kPatternLevelRight);
       }
-      if (partialDigit && pos < numDigits) {
-        mPatternWriter.writePatternAt(pos++, kPatternLevelLeft);
+      if (partialDigit && mPatternWriter.pos() < numDigits) {
+        mPatternWriter.writePattern(kPatternLevelLeft);
       }
-      mPatternWriter.clearToEnd(pos);
+      mPatternWriter.clearToEnd();
     }
 
   private:
@@ -89,7 +90,7 @@ class LevelWriter {
     LevelWriter& operator=(const LevelWriter&) = delete;
 
   private:
-    PatternWriter<T_LED_MODULE> mPatternWriter;
+    PatternWriter<T_LED_MODULE>& mPatternWriter;
 };
 
 }