+++ src/ace_segment_writer/NumberWriter.h
@@ -31,40 +31,39 @@ SOFTWARE.
 
 namespace ace_segment {
 
-/** Total number of characters in the kHexCharPatterns[] set. */
-const uint8_t kNumHexCharPatterns = 18;
+/** Total number of characters in the kDigitPatterns[] set. */
+const uint8_t kNumDigitPatterns = 18;
 
 /** Bit pattern map for hex characters. */
-extern const uint8_t kHexCharPatterns[kNumHexCharPatterns];
+extern const uint8_t kDigitPatterns[kNumDigitPatterns];
 
 /**
  * The type of the character set supported by many methods in the NumberWriter
- * class usually containing the string "HexChar" in its name. This custom
+ * class usually containing the string "Digit" in its name. This custom
  * character set is not ASCII to save flash memory. It is a restricted set that
  * starts with 0 and goes up to 0xF to support hexadecimal digits. In addition,
  * the character set adds few more characters for convenience:
  *
- *  * kHexCharSpace = 0x10 = space character
- *  * kHexCharMinus = 0x11 = minus character
+ *  * kDigitSpace = 0x10 = space character
+ *  * kDigitMinus = 0x11 = minus character
  *
- * The `hexchar_t` typedef is an alias for `uint8_t` and unfortunately, C++
- * will not prevent mixing a normal `char` or `uint8_t` with a `hexchar_t`.
- * However, it does make the documentation of the various methods more
- * self-explanatory.
+ * The `digit_t` typedef is an alias for `uint8_t`. Unfortunately, C++ will not
+ * prevent mixing a normal `char` or `uint8_t` with a `digit_t`. However, it
+ * does make the documentation of the various methods more self-explanatory.
  */
-typedef uint8_t hexchar_t;
+typedef uint8_t digit_t;
 
-/** The code point in kHexCharPatterns[] corresponding to a space character. */
-const hexchar_t kHexCharSpace = 0x10;
+/** The code point in kDigitPatterns[] corresponding to a space character. */
+const digit_t kDigitSpace = 16;
 
-/** The code point in kHexCharPatterns[] corresponding to a minus character. */
-const hexchar_t kHexCharMinus = 0x11;
+/** The code point in kDigitPatterns[] corresponding to a minus character. */
+const digit_t kDigitMinus = 17;
 
 /**
  * The NumberWriter supports converting decimal and hexadecimal numbers to
  * segment patterns expected by LedModule. The character set includes 0 to F,
- * and a few other characters which should be self-explanatory: kHexCharSpace
- * and kHexCharMinus.
+ * and a few other characters which should be self-explanatory: kDigitSpace
+ * and kDigitMinus.
  *
  * @tparam T_LED_MODULE the class of the underlying LED Module, often LedModule
  *    but other classes with the same generic public methods can be substituted
@@ -73,8 +72,8 @@ template <typename T_LED_MODULE>
 class NumberWriter {
   public:
     /** Constructor. */
-    explicit NumberWriter(T_LED_MODULE& ledModule) :
-        mPatternWriter(ledModule)
+    explicit NumberWriter(PatternWriter<T_LED_MODULE>& patternWriter) :
+        mPatternWriter(patternWriter)
     {}
 
     /** Get the underlying LedModule. */
@@ -83,43 +82,23 @@ class NumberWriter {
     /** Get the underlying PatternWriter. */
     PatternWriter<T_LED_MODULE>& patternWriter() { return mPatternWriter; }
 
-    /**
-     * Write the hex character `c` at position `pos`. If `c` falls outside the
-     * valid range of the kHexCharPatterns set, a `kHexCharSpace` character is
-     * printed instead.
-     */
-    void writeHexCharAt(uint8_t pos, hexchar_t c) {
-      writeInternalHexCharAt(
-          pos, ((uint8_t) c < kNumHexCharPatterns) ? c : kHexCharSpace);
-    }
+    /** Reset cursor to home. */
+    void home() { mPatternWriter.home(); }
 
     /**
-     * Write 2 hex characters given in the arguments. A convenience function
-     * that seems to occur frequently when writing time or date components.
+     * Write the digit `c` at position `pos`. If `c` falls outside the valid
+     * range of the kDigitPatterns set, a `kDigitSpace` character is printed
+     * instead.
      */
-    void writeHexChars2At(uint8_t pos, hexchar_t c, hexchar_t d) {
-      writeHexCharAt(pos++ ,c);
-      writeHexCharAt(pos, d);
+    void writeDigit(digit_t c) {
+      write(((uint8_t) c < kNumDigitPatterns) ? c : kDigitSpace);
     }
 
     /** Write the `len` hex characters given by `s` at LED `pos`. */
-    void writeHexCharsAt(uint8_t pos, const hexchar_t s[], uint8_t len) {
+    void writeDigits(const digit_t s[], uint8_t len) {
       for (uint8_t i = 0; i < len; ++i) {
-        writeHexCharAt(pos++, s[i]);
+        writeDigit(s[i]);
       }
-    }
-
-    /**
-     * Write a 2-digit BCD number at position, which involves just printing the
-     * number as a hexadecimal number. For example, 0x12 is printed as "12", but
-     * 0x1A is printed as "1 ".
-     */
-    void writeBcd2At(uint8_t pos, uint8_t bcd) {
-      uint8_t high = (bcd & 0xF0) >> 4;
-      uint8_t low = (bcd & 0x0F);
-      if (high > 9) high = kHexCharSpace;
-      if (low > 9) low = kHexCharSpace;
-      writeHexChars2At(pos, high, low);
     }
 
     /**
@@ -128,23 +107,22 @@ class NumberWriter {
      * than 100, then print "  " (2 spaces). Useful for month, day, hour,
      * minute, second.
      *
-     * @param pos LED digit position
      * @param d decimal value to write
      * @param padPattern leading pattern, usually kPatternSpace or kPattern0
      */
-    void writeDec2At(uint8_t pos, uint8_t d, uint8_t padPattern = kPattern0) {
+    void writeDec2(uint8_t d, uint8_t padPattern = kPattern0) {
       if (d >= 100) {
-        mPatternWriter.writePatternAt(pos++, kPatternSpace);
-        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+        mPatternWriter.writePattern(kPatternSpace);
+        mPatternWriter.writePattern(kPatternSpace);
       } else {
         uint8_t tens = d / 10;
         uint8_t ones = d - 10 * tens;
         if (tens == 0) {
-          mPatternWriter.writePatternAt(pos++, padPattern);
+          mPatternWriter.writePattern(padPattern);
         } else {
-          writeHexCharAt(pos++, tens);
+          writeDigit(tens);
         }
-        writeHexCharAt(pos, ones);
+        writeDigit(ones);
       }
     }
 
@@ -153,52 +131,60 @@ class NumberWriter {
      * with a the `padPattern` (default kPattern0). If the number is greater
      * than 10000, print 4 spaces.
      */
-    void writeDec4At(uint8_t pos, uint16_t dd, uint8_t padPattern = kPattern0) {
+    void writeDec4(uint16_t dd, uint8_t padPattern = kPattern0) {
       if (dd >= 10000) {
-        mPatternWriter.writePatternAt(pos++, kPatternSpace);
-        mPatternWriter.writePatternAt(pos++, kPatternSpace);
-        mPatternWriter.writePatternAt(pos++, kPatternSpace);
-        mPatternWriter.writePatternAt(pos++, kPatternSpace);
+        mPatternWriter.writePattern(kPatternSpace);
+        mPatternWriter.writePattern(kPatternSpace);
+        mPatternWriter.writePattern(kPatternSpace);
+        mPatternWriter.writePattern(kPatternSpace);
       } else {
         uint8_t high = dd / 100;
         uint8_t low = dd - high * 100;
         if (high == 0) {
-          mPatternWriter.writePatternAt(pos++, padPattern);
-          mPatternWriter.writePatternAt(pos++, padPattern);
-          writeDec2At(pos, low, padPattern);
+          mPatternWriter.writePattern(padPattern);
+          mPatternWriter.writePattern(padPattern);
+          writeDec2(low, padPattern);
         } else {
-          writeDec2At(pos, high, padPattern);
-          pos += 2;
-          writeDec2At(pos, low, kPattern0);
+          writeDec2(high, padPattern);
+          writeDec2(low, kPattern0);
         }
       }
     }
 
-    /** Write the 2-digit (8-bit) hexadecimal byte 'b' at pos. */
-    void writeHexByteAt(uint8_t pos, uint8_t b) {
-      uint8_t low = (b & 0x0F);
-      b >>= 4;
-      uint8_t high = (b & 0x0F);
+    /**
+     * Write a 2-digit BCD number at position, which involves just printing the
+     * number as a hexadecimal number. For example, 0x12 is printed as "12", but
+     * 0x1A is printed as "1 ".
+     */
+    void writeBcd(uint8_t bcd) {
+      uint8_t high = (bcd & 0xF0) >> 4;
+      uint8_t low = (bcd & 0x0F);
+      if (high > 9) high = kDigitSpace;
+      if (low > 9) low = kDigitSpace;
+      writeDigit(high);
+      writeDigit(low);
+    }
 
-      // We can use writeInternalHexCharAt() because low and high are guaranteed
-      // to be a validate hex character.
-      writeInternalHexCharAt(pos++, high);
-      writeInternalHexCharAt(pos++, low);
+    /** Write the 2-digit (8-bit) hexadecimal byte 'b' at pos. */
+    void writeHexByte(uint8_t b) {
+      uint8_t low = (b & 0x0F);
+      uint8_t high = ((b >> 4) & 0x0F);
+      write(high);
+      write(low);
     }
 
     /** Write the 4 digit (16-bit) hexadecimal word at pos. */
-    void writeHexWordAt(uint8_t pos, uint16_t w) {
+    void writeHexWord(uint16_t w) {
       uint8_t low = (w & 0xFF);
       uint8_t high = (w >> 8) & 0xFF;
-      writeHexByteAt(pos, high);
-      writeHexByteAt(pos + 2, low);
+      writeHexByte(high);
+      writeHexByte(low);
     }
 
     /**
      * Write the 16-bit unsigned number `num` as a decimal number at pos. The
      * lead padding character is always kPatternSpace.
      *
-     * @param pos start position of the number
      * @param num unsigned decimal number, 0-65535
      * @param boxSize size of box. This is meant to be similar to the "%-5d" or
      *    "%5d" specifier to the printf() function. The actual digits may spill
@@ -210,74 +196,63 @@ class NumberWriter {
      * @return number of characters actually written, even if the characters
      *    bled over the end of the LED segments
      */
-    uint8_t writeUnsignedDecimalAt(
-        uint8_t pos,
-        uint16_t num,
-        int8_t boxSize = 0) {
-
+    uint8_t writeUnsignedDecimal(uint16_t num, int8_t boxSize = 0) {
       const uint8_t bufSize = 5;
-      hexchar_t buf[bufSize];
+      digit_t buf[bufSize];
       uint8_t start = toDecimal(num, buf, bufSize);
 
-      return writeHexCharsInsideBoxAt(
-          pos, &buf[start], bufSize - start, boxSize);
+      return writeDigitsInsideBox(&buf[start], bufSize - start, boxSize);
     }
 
-    /** Same as writeUnsignedDecimalAt() but prepends a '-' sign if negative. */
-    uint8_t writeSignedDecimalAt(
-        uint8_t pos,
-        int16_t num,
-        int8_t boxSize = 0) {
-
+    /** Same as writeUnsignedDecimal() but prepends a '-' sign if negative. */
+    uint8_t writeSignedDecimal(int16_t num, int8_t boxSize = 0) {
       // Even -32768 turns into +32768, which is exactly what we want
       bool negative = num < 0;
       uint16_t absNum = negative ? -num : num;
 
       const uint8_t bufSize = 6;
-      hexchar_t buf[bufSize];
+      digit_t buf[bufSize];
       uint8_t start = toDecimal(absNum, buf, bufSize);
       if (negative) {
-        buf[--start] = kHexCharMinus;
+        buf[--start] = kDigitMinus;
       }
 
-      return writeHexCharsInsideBoxAt(
-          pos, &buf[start], bufSize - start, boxSize);
+      return writeDigitsInsideBox(&buf[start], bufSize - start, boxSize);
     }
 
     /**
      * Write the decimal point for the pos. Clock LED modules will attach the
      * colon segment to one of the decimal points.
      */
-    void writeDecimalPointAt(uint8_t pos, bool state = true) {
-      mPatternWriter.writeDecimalPointAt(pos, state);
+    void setDecimalPointAt(uint8_t pos, bool state = true) {
+      mPatternWriter.setDecimalPointAt(pos, state);
     }
 
     /** Clear the entire display. */
-    void clear() { clearToEnd(0); }
+    void clear() { mPatternWriter.clear(); }
 
     /**
-     * Clear the display from `pos` to the end. Useful after calling
-     * writeSignedDecimalAt() and writeUnsignedDecimalAt() which prints a
+     * Clear the display from the current position to the end. Useful after
+     * calling writeSignedDecimal() and writeUnsignedDecimal() which prints a
      * variable number of digits.
      */
-    void clearToEnd(uint8_t pos) { mPatternWriter.clearToEnd(pos); }
+    void clearToEnd() { mPatternWriter.clearToEnd(); }
 
   private:
     // disable copy-constructor and assignment operator
     NumberWriter(const NumberWriter&) = delete;
     NumberWriter& operator=(const NumberWriter&) = delete;
 
-    /** Similar to writeHexCharAt() without performing bounds check. */
-    void writeInternalHexCharAt(uint8_t pos, hexchar_t c) {
-      uint8_t pattern = pgm_read_byte(&kHexCharPatterns[(uint8_t) c]);
-      mPatternWriter.writePatternAt(pos, pattern);
+    /** Internal version of writeDigit() without performing bounds check. */
+    void write(digit_t c) {
+      uint8_t pattern = pgm_read_byte(&kDigitPatterns[(uint8_t) c]);
+      mPatternWriter.writePattern(pattern);
     }
 
-    /** Similar to writeHexCharsAt() without performing bounds check. */
-    void writeInternalHexCharsAt(uint8_t pos, const hexchar_t s[],
-        uint8_t len) {
+    /** Similar to write() without performing bounds check. */
+    void writeInternalDigits(const digit_t s[], uint8_t len) {
       for (uint8_t i = 0; i < len; ++i) {
-        writeInternalHexCharAt(pos++, s[i]);
+        write(s[i]);
       }
     }
 
@@ -288,17 +263,14 @@ class NumberWriter {
      *
      * @param boxSize if negative, left justified; if postive, right justified
      */
-    uint8_t writeHexCharsInsideBoxAt(
-        uint8_t pos,
-        const hexchar_t s[],
-        uint8_t len,
-        int8_t boxSize) {
+    uint8_t writeDigitsInsideBox(
+        const digit_t s[], uint8_t len, int8_t boxSize) {
 
       uint8_t absBoxSize = (boxSize < 0) ? -boxSize : boxSize;
 
       // if the box is too small, print normally
       if (len >= absBoxSize) {
-        writeInternalHexCharsAt(pos, s, len);
+        writeInternalDigits(s, len);
         return len;
       }
 
@@ -306,20 +278,19 @@ class NumberWriter {
       uint8_t padSize = absBoxSize - len;
       if (boxSize < 0) {
         // left justified
-        writeInternalHexCharsAt(pos, s, len);
-        pos += len;
-        while (padSize--) writeInternalHexCharAt(pos++, kHexCharSpace);
+        writeInternalDigits(s, len);
+        while (padSize--) write(kDigitSpace);
       } else {
         // right justified
-        while (padSize--) writeInternalHexCharAt(pos++, kHexCharSpace);
-        writeInternalHexCharsAt(pos, s, len);
+        while (padSize--) write(kDigitSpace);
+        writeInternalDigits(s, len);
       }
 
       return absBoxSize;
     }
 
     /**
-     * Convert the integer num to an array of HexChar in the provided buf, with
+     * Convert the integer num to an array of Digit in the provided buf, with
      * the least significant digit going to buf[bufSize-1], and then working
      * backwards to the most significant digit.
      *
@@ -331,7 +302,7 @@ class NumberWriter {
      * @return index into buf that points to the start of the converted number,
      * e.g. for a single digit number, the returned value will be `bufSize-1`.
      */
-    static uint8_t toDecimal(uint16_t num, hexchar_t buf[], uint8_t bufSize) {
+    static uint8_t toDecimal(uint16_t num, digit_t buf[], uint8_t bufSize) {
       uint8_t pos = bufSize;
       while (true) {
         if (num < 10) {
@@ -346,7 +317,7 @@ class NumberWriter {
     }
 
   private:
-    PatternWriter<T_LED_MODULE> mPatternWriter;
+    PatternWriter<T_LED_MODULE> &mPatternWriter;
 };
 
 }